responsible-ai-music-scraper-proto-main/cli_data_scraper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "responsible_data_scraper"
version = "1.0.0"
authors = [
    { name = "Winand Metz", email = "winand.metz@hva.nl" },
    { name = "Caleigh Esterhuyse", email = "caleigh.esterhuyse@hva.nl" },
    { name = "Stef Kan", email = "stef.kan@hva.nl"},
    { name = "Max Fijma", email = "max.fijma@hva.nl"}
]
description = "A scraper for the Responsible AI Music project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.2",
    "requests>=2.30.0",
    "lyricsgenius>=3.0.1",
    "ftfy>=6.1.1",
    "lingua-language-detector>=1.3.2"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.hatch.build.targets.sdist.force-include]
"bin/data_scraper.py" = "data_collector/data_scraper.py"

[tool.hatch.build.targets.wheel.force-include]
"bin/data_scraper.py" = "data_collector/data_scraper.py"

[project.scripts]
data_scraper = "data_collector.data_scraper:main"

[project.urls]
"Source" = "https://gitlab.fdmci.hva.nl/metzwm/responsible-ai-music-scraper-proto"
"Bug Tracker" = "https://gitlab.fdmci.hva.nl/metzwm/responsible-ai-music-scraper-proto/-/issues"

[tool.hatch.build]
ignore-vcs = true
